--- a/gradle.properties
+++ b/gradle.properties
@@ -1,5 +1,5 @@
 org.gradle.caching=true
 org.gradle.jvmargs='-Dfile.encoding=UTF-8'
 group=org.ballerinax.datamapper
-version=2.0.0-SNAPSHOT
+version=2.0.0-Preview2-SNAPSHOT
 systemProp.org.gradle.internal.publish.checksums.insecure=true