## Purpose
A bad sad error was thrown when building a project which imported an external connector module into a Ballerina app using Ballerina Swan Lake Preview 1. Specifically, this error is thrown when the build process was invoked for the second round. This is because when the compilation was done for the first time, the data mapper compiler extension generates functions JSON file and when the compilation is run again it fails to invoke the data mapper compiler extension because the functions JSON file already exists.

## Goals
Fix bug which is caused when importing external connector modules.

## Approach
Add a code snippet to check whether the processing happens within the same source project or not.

## User stories
N/A

## Release note
N/A

## Documentation
N/A

## Training
N/A

## Certification
N/A

## Marketing
N/A

## Automation tests
N/A

## Performance tests
A performance test was carried out on a computer having 4 CPU cores, and 16GB RAM with more than 100GB free disk space. The computer was installed Ubuntu 18.04, JDK 1.8. Tests were carried out using three Ballerina connectors and two Ballerina by example sample applications.

| Connector/Application  |  Latency with Data Mapper Plugin (s) | Latency without Data Mapaper Plugin (s)  | Increase in Latency (s)  | Percentage Increase in build time (%)  |
|---|---|---|---|---|
| Slack | 3.49 | 3.33 | 0.16 | 4.95|
|Github | 3.83 | 3.70 | 0.13 | 3.51|
|Sfdc | 4.03 | 3.87 | 0.16 | 4.03|
|Workers example from BBE | 2.48 | 2.55 | 0.04 | 1.58|
|JSON arrays example from BBE | 2.463413834 | 2.456724361 | 0.0066 | 0.27|



## Security checks
 - Followed secure coding standards in http://wso2.com/technical-reports/wso2-secure-engineering-guidelines? yes
 - Ran FindSecurityBugs plugin and verified report? no
 - Confirmed that this PR doesn't commit any keys, passwords, tokens, usernames, or other secrets? yes

## Samples
N/A

## Related PRs
[Need a flag to indicate a packageNode is from the source project](https://github.com/ballerina-platform/ballerina-lang/issues/24441)

## Migrations (if applicable)
N/A

## Test environment
JDK 1.8, Ubuntu 18.04, Ballerina Swan Lake Preview 1
 
## Learning
N/A